Add Game.valid_move? and some test cases
I realized that there's an entire mechanism that I don't
support: Flipping the available moves when black is playing.

If I have a card that allows me to move {0, 1}, that is, one space
"up" the board, that's fine.  But if it's black's turn, that card
actually provides them the move {0, -1}, since that's "up" from
black's perspective.  Going to need to start looking at
the active player in a bunch of these functions.