Initial implementation of AI (min-max w/ alpha beta pruning)
The algorithm is sound, but this code is rough and needs to be cleaned up.

Eventually I would like to have a simple, easy "Play against the computer"
mode in the UI.  Right now, there is a rough "Have the AI choose the next move"
button in the game.

AI plays well, but:
  - When it knows that it is going to lose no matter what,
  it returns 'nil' as the move and the "AI Move" button will temporarily
  crash the game.

  - It could compute some moves in parallel  (Not trivial, needs special rules.  See: http://www.netlib.org/utk/lsi/pcwLSI/text/node351.html for guidance in parallelizing it)

  - Need some tests and possibly better code organization