Add/remove cards
I felt the sideways only cards were really lame.  The games felt
ruined if I drew one to start with.

I also felt the "Wild Pig" card is so core, there needs to be more of
it. So I added a copy of it called piglet, and added a derivative that
can also go backwards.